server/requirements.txt
channels-redis==2.3.3
djangorestframework==3.9.2
nose==1.3.7
Pillow==6.0.0
pytest-asyncio==0.10.0
pytest-django==3.4.8
asgiref==2.1
daphne==2.1
channels==2.1